+++ pages/dashboardAdmin/hapus_histori.php
@@ -1,6 +1,23 @@
 <?php
+session_start();
 require_once "../../config/app.php";
 $id_history = (int)$_GET['id_history'];
+
+if (!isset($_SESSION['adminLogin'])) {
+    echo "<script>
+    alert('Anda harus login admin terlebih dahulu!');
+    document.location.href='../../index.php';
+     </script>";
+    exit;
+}
+
+if ($_SESSION['adminRole'] != 1) {
+    echo "<script>
+    alert('Role anda tidak sesuai!');
+    document.location.href='index.php';
+     </script>";
+    exit;
+}
 
 if (hapus_histori($id_history) > 0) {
     echo "<script>
